% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkDirCoo}
\alias{checkDirCoo}
\title{make sure all coordinates of outlines stored in an object of class "Coo" are oriented coherently}
\usage{
checkDirCoo(Coo)
}
\arguments{
\item{Coo}{object of class "Coo"}
}
\value{
input with outlines oriented coherently
}
\description{
make sure all coordinates of outlines stored in an object of class "Coo" are oriented coherently
}

