% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fixalotMissing.r
\name{checkmatArray}
\alias{checkmatArray}
\title{check if there are pairwise complementary data missing}
\usage{
checkmatArray(nalist)
}
\description{
check if there are pairwise complementary data missing
}

