% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwdcv.r
\name{dwdcv}
\alias{dwdcv}
\title{leaving one out cross-validation for DWD}
\usage{
dwdcv(x, y, ...)
}
\arguments{
\item{x}{predictor}

\item{y}{binary coded variable for classification}

\item{...}{arguments passed to kerndwd and predict.kerndwd}
}
\value{
vector of cross-validated predictions
}
\description{
leaving one out cross-validation for DWD
}

