\name{ef2coo}
\alias{ef2coo}
\title{convert harmonics to 2D-coordinates}
\usage{
ef2coo(x, nh = NULL, nb.pts = 300)
}
\arguments{
  \item{x}{vector, matrix or Coe object}

  \item{nh}{integer (vector): number of harmonics to use}

  \item{nb.pts}{number of coordinats}
}
\value{
matrix with 2D coordinates
}
\description{
convert harmonics to 2D-coordinates
}

