% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getEllipse.r
\name{getEllipse}
\alias{getEllipse}
\title{compute ellipse fitting four points as the extrema}
\usage{
getEllipse(points, npt)
}
\arguments{
\item{points}{4 x 3 matrix with rows containing the extrema of an ellipse in clockwise or counterclockwise fashion}

\item{npt}{integer: amount of points on the ellipse to generate}
}
\value{
returns a matrix with points showing an allipse
}
\description{
compute ellipse fitting four points as the extrema
}

