% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fixalotMissing.r
\name{getNAlistBool}
\alias{getNAlistBool}
\title{get a list with all rows from an nalist (output from nafunArray) that are actually NAs}
\usage{
getNAlistBool(nalist)
}
\description{
get a list with all rows from an nalist (output from nafunArray) that are actually NAs
}

