% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EFAhelper.r
\name{getStart}
\alias{getStart}
\title{get index of starting point given a close estimate}
\usage{
getStart(lms, coords)
}
\arguments{
\item{lms}{list of starting points (as 1 x 2 matrices)}

\item{coords}{object of class Coo}
}
\value{
vector containing indices of starting points
}
\description{
get index of starting point given a close estimate determined by setLM
}

