\name{ggMatvsVar}
\alias{ggMatvsVar}
\title{create Plot for each column of PCs against var}
\usage{
  ggMatvsVar(PCs, var, xlab = xname, ncol = 2,
    group = NULL, layout = NULL, legend.title = NULL)
}
\arguments{
  \item{PCs}{Matrix for y-Axis}

  \item{var}{vector containing data for x-axis}

  \item{xlab}{x-axis label}

  \item{ncol}{layout columns}

  \item{group}{grouping variable}

  \item{layout}{layout matrix}

  \item{legend.title}{custom legend title}
}
\description{
  create Plot for each column of PCs against var
}

