% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fixalotMissing.r
\name{nafunArray}
\alias{nafunArray}
\title{for a 3D-array get a list of vectors indicating lines with NAs 0=data, 1=NA}
\usage{
nafunArray(x)
}
\description{
for a 3D-array get a list of vectors indicating lines with NAs 0=data, 1=NA
}

