% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EFAhelper.r
\name{normEFA}
\alias{normEFA}
\title{run efourier_norm on object of class Coe}
\usage{
normEFA(efa, output = "size")
}
\arguments{
\item{efa}{object of class Coe}

\item{output}{if "size" only a list with the size of the first ellipse is returned, the complete output of efourier_norm otherwise.}
}
\description{
run efourier_norm on object of class Coe and report values
}

