\name{plotCurves}
\alias{plotCurves}
\title{plot Curves generated wiht setCurvature}
\usage{
plotCurves(curve, text = FALSE, color = "rainbow")
}
\arguments{
  \item{curve}{object read by readCurves}

  \item{text}{name curves}

  \item{color}{color curves}
}
\description{
plot Curves generated wiht setCurvature
}

