% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EFAhelper.r
\name{resortCoord}
\alias{resortCoord}
\title{resort coordinates based on indices of starting points}
\usage{
resortCoord(coords, ind)
}
\arguments{
\item{coords}{object of class Coo}

\item{vector}{of indices}
}
\value{
Coo object with resorted coordinates
}
\description{
resort coordinates based on indices of starting points (determined by getStart)
}

