% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPHARM.r
\name{spharmParaToMesh}
\alias{spharmParaToMesh}
\title{a wrapper for ParaToSPHARMMeshCLP}
\usage{
spharmParaToMesh(paraname, surfname, args = NULL, outname = "default_",
  helpargs = FALSE, readvtk = TRUE)
}
\description{
a wrapper for ParaToSPHARMMeshCLP
}

