% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cholMaha}
\alias{cholMaha}
\title{pairwise mahalnobis distance}
\usage{
cholMaha(X)
}
\arguments{
\item{x}{data matrix}
}
\value{
pairwise squared mahalanobis distances
}
\description{
pairwise mahalnobis distance
}
\note{
copied from \url{http://stats.stackexchange.com/a/81710}
}

