\name{fixitall}
\alias{fixitall}
\title{run all functions above and select the best version of each specimen (i.e. with the least missing landmarks}
\usage{
fixitall(bilatfix)
}
\description{
run all functions above and select the best version of each
specimen (i.e. with the least missing landmarks
}

