% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixalotMissing.r
\name{fixitall}
\alias{fixitall}
\title{run all functions above and select the best version of each specimen (i.e. with the least missing landmarks}
\usage{
fixitall(bilatfix)
}
\description{
run all functions above and select the best version of each specimen (i.e. with the least missing landmarks
}

