% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNAlistBool}
\alias{getNAlistBool}
\title{get a list with all rows from an nalist (output from nafunArray) that are actually NAs}
\usage{
getNAlistBool(nalist)
}
\description{
get a list with all rows from an nalist (output from nafunArray) that are actually NAs
}

