% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggMatvsVar}
\alias{ggMatvsVar}
\alias{ggMatvsVar2}
\title{create Plot for each column of PCs against var}
\usage{
ggMatvsVar(PCs, var, xlab = xname, ncol = 2, group = NULL,
  layout = NULL, legend.title = NULL, pt.size = 1)

ggMatvsVar2(PCs, var, xlab = xname, ncol = 2, group = NULL,
  layout = NULL, legend.title = NULL, pt.size = 1, lwd = 1)
}
\arguments{
\item{PCs}{Matrix for y-Axis}

\item{var}{vector containing data for x-axis}

\item{xlab}{x-axis label}

\item{ncol}{layout columns}

\item{group}{grouping variable}

\item{layout}{layout matrix}

\item{legend.title}{custom legend title}

\item{pt.size}{point size}

\item{lwd}{line width of regression line}
}
\description{
create Plot for each column of PCs against var
}

