\name{pairwiseTPS}
\alias{pairwiseTPS}
\title{based on an array and the result from getNAlistBool map each specimen onto all others using TPS}
\usage{
pairwiseTPS(bilatfix, nalistBool)
}
\description{
based on an array and the result from getNAlistBool map
each specimen onto all others using TPS
}

