% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crestlines.r
\name{plotCurves}
\alias{plotCurves}
\title{plot Curves generated wiht setCurvature}
\usage{
plotCurves(curve, text = FALSE, color = "rainbow")
}
\arguments{
\item{curve}{object read by readCurves}

\item{text}{name curves}

\item{color}{color curves}
}
\description{
plot Curves generated wiht setCurvature
}

