% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFAhelper.r
\name{setLM}
\alias{setLM}
\title{set starting point in a list of outlines}
\usage{
setLM(x)
}
\arguments{
\item{x}{list of coordinate matrices}
}
\value{
a list of coordinates close to the starting point
}
\description{
set starting point in a list of outlines
}

