% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPHARM.r
\name{spharmCompute}
\alias{spharmCompute}
\title{A simple wrapper for the command line tools of SPHARM-PDM}
\usage{
spharmCompute(x, imagename, imagetype = "mha", clean = TRUE,
  flipTemplate = NULL, spacing = rep(1, 3), outfolder = "./",
  iter = 100, subdivLevel = 20, GenArgs = NULL, ParaArgs = NULL,
  readvtk = TRUE, silent = FALSE)
}
\description{
A simple wrapper for the command line tools of SPHARM-PDM, starting from an object of class mesh3d
}

